# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Global Tongue API Documentation",
        "description": "Heirarchy of API | Courses -> Levels -> Units -> Lessons -> Exercises",
        "contact": {
            "email": "gemeelijah@gmail.com"
        },
        "license": {
            "name": "www.globaltongueedu.com",
            "url": "https://globaltongueedu.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/public/api/units": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "Get Unit List",
                "description": "Get Unit List as Array",
                "operationId": "UnitIndex",
                "responses": {
                    "200": {
                        "description": "Get Unit List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/public/api/units/{id}": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "Show Unit Details",
                "description": "Show Unit Details",
                "operationId": "showUnit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Unit Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/public/api/units/by/{course_id}/{level_id}": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "Get Units by Parents ID",
                "description": "Get list of units associated with a specific course",
                "operationId": "getUnitsByCourseID",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "level_id",
                        "in": "path",
                        "description": "ID of the level",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Units for the specified Parents ID"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "No units found for the specified Parents ID"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}